Allow eager loading associations w/ a renamed fk
Because Korma's :transforms/:prepares features allow users to rename
fields, we can't necessarily assume that the foreign key passed to the
database query is the same as the one in the result set. This removes
that assumption from the eager-loading stitcher by transforming the
foreign key in the same way as records are transformed.